+++ aim/engine/configs.py
@@ -1,3 +1,4 @@
+# aim
 AIM_REPO_NAME = '.aim'
 AIM_OBJECTS_DIR_NAME = 'objects'
 AIM_METRICS_DIR_NAME = 'metrics'
@@ -5,13 +6,26 @@ AIM_ANNOT_DIR_NAME = 'annotations'
 AIM_MEDIA_DIR_NAME = 'media'
 AIM_IMAGES_DIR_NAME = 'images'
 AIM_MODELS_DIR_NAME = 'models'
+AIM_CORR_DIR_NAME = 'correlations'
+AIM_CORR_DIRS_NAME = 'dirs'
 AIM_CONFIG_FILE_NAME = 'config.json'
+AIM_DIFF_DIR_NAME = 'diff'
+AIM_DIFF_FILE_NAME = 'diff.txt'
+AIM_PARAMS_DIR_NAME = 'hyperparameters'
+
+# Branches
 AIM_DEFAULT_BRANCH_NAME = 'default'
 
+# Commits
+AIM_COMMIT_INDEX_DIR_NAME = 'index'
+AIM_COMMIT_CONFIG_FILE_NAME = 'config.json'
+
+# Logger
 AIM_LOGGING_DIR_NAME = 'logs'
 AIM_CLI_ACTIVITIED_LOG = 'activities.log'
 AIM_SDK_ERROR_LOG = 'errors.log'
 
+# aim Profile
 AIM_PROFILE_NAME = '.aim_profile'
 AIM_PROFILE_CONFIG_FILE_NAME = 'config.json'
 AIM_PROFILE_SSH_DIR_NAME = '.ssh'