+++ aim/sdk/artifacts/__init__.py
@@ -2,3 +2,6 @@ from aim.sdk.artifacts.metric import *
 from aim.sdk.artifacts.media import *
 from aim.sdk.artifacts.misclassification import *
 from aim.sdk.artifacts.model import *
+from aim.sdk.artifacts.correlation import *
+from aim.sdk.artifacts.distribution import *
+from aim.sdk.artifacts.hyperparameters import *