+++ examples/pytorch_track_params.py
@@ -1,5 +1,7 @@
 import aim
 from aim import track
+import random
+import math
 
 import torch
 import torch.nn as nn
@@ -14,6 +16,14 @@ num_epochs = 5
 num_classes = 10
 batch_size = 50
 learning_rate = 0.01
+
+# aim - Track hyper parameters
+track(aim.hyperparams, {
+    'num_epochs': num_epochs,
+    'num_classes': num_classes ,
+    'batch_size': batch_size ,
+    'learning_rate': learning_rate,
+})
 
 # MNIST dataset
 train_dataset = torchvision.datasets.MNIST(root='./data/',
@@ -89,54 +99,11 @@ for epoch in range(num_epochs):
                   'Loss: {:.4f}'.format(epoch + 1, num_epochs, i + 1,
                                         total_step, loss.item()))
 
-            # aim - Track model loss function
-            track(aim.loss, 'loss', loss.item())
-
-            # aim - Track last layer correlation
-            track(aim.label_correlation, 'corr', outputs, labels=[
-                0, 1, 2, 3, 4, 5, 6, 7, 8, 9,
-            ])
-
-            # aim - Track model weights and gradients
-            track(aim.weights, model)
-            track(aim.gradients, model)
-
-        for l in range(len(labels)):
-            for o in range(len(outputs)):
-                if labels[l].item() == outputs[o].argmax().item() and l != o:
-                    # count fp
-                    false_positives += 1
-
-        for l in range(len(labels)):
-            if labels[l].item() != outputs[l].argmax().item():
-                # count fn
-                false_negatives += 1
-
-                # track images
-                if epoch >= 2 and saved_img < 50:
-                    saved_img += 1
-                    # aim - Track misclassified images
-                    img = track(aim.image, images[l])
-                    track(aim.misclassification, 'miscls', img,
-                          labels[l].item(),
-                          outputs[l].argmax().item())
-
-    learning_rate /= 2
-
-    # aim - Track model checkpoints
-    track(aim.checkpoint,
-          'checkpoint_test', 'chp_epoch_{}'.format(epoch),
-          model, epoch, lr_rate=learning_rate,
-          meta={
-              'learning_rate': learning_rate,
-              'false_positives': false_positives,
-              'false_negatives': false_negatives,
-              'drop_out': 0.5,
-              'batch_size': 10,
-              'kernel_size': 2,
-              'stride': 2,
-          })
-
+            correct = 0
+            total = 0
+            _, predicted = torch.max(outputs.data, 1)
+            total += labels.size(0)
+            correct += (predicted == labels).sum().item()
 # Test the model
 model.eval()
 with torch.no_grad():